+++ pack_o_daemon/run.py
@@ -34,3 +34,6 @@ def main():
         
         sys.exit()
     app.MainLoop()
+
+if __name__ == '__main__':
+    main()