+++ run.sh
@@ -1,0 +1,3 @@
+python setup.py sdist bdist_wheel
+pip install --upgrade .
+pack_o_daemon